refactor/#28: strategy 및 guard의 역할에 맞게 코드 수정, 회원가입, 로그인을 auth 모듈에서 처리하도록 수정
