+++ src/apis/apis.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
+import { AuthModule } from '@src/apis/auth/auth.module';
 import { PostModule } from '@src/apis/post/post.module';
-import { TokenModule } from '@src/apis/token/token.module';
 import { UserModule } from '@src/apis/user/user.module';
 
 @Module({
-  imports: [TokenModule, PostModule, UserModule],
+  imports: [AuthModule, PostModule, UserModule],
 })
 export class ApisModule {}
