+++ src/apis/auth/commands/generate-access-token/generate-access-token.command.ts
@@ -1,13 +1,14 @@
 import { ICommand } from '@nestjs/cqrs';
 import { Command, CommandProps } from '@src/libs/ddd/command.base';
-import { AggregateID } from '@src/libs/ddd/entity.base';
 
 export class GenerateAccessTokenCommand extends Command implements ICommand {
-  readonly userId: AggregateID;
+  readonly email: string;
+  readonly password: string;
 
   constructor(props: CommandProps<GenerateAccessTokenCommand>) {
     super(props);
 
-    this.userId = props.userId;
+    this.email = props.email;
+    this.password = props.password;
   }
 }
