+++ src/apis/user/controllers/user.swagger.ts
@@ -1,7 +1,6 @@
 import { applyDecorators, HttpStatus } from '@nestjs/common';
 import {
   ApiBearerAuth,
-  ApiCreatedResponse,
   ApiNoContentResponse,
   ApiOkResponse,
   ApiOperation,
@@ -9,7 +8,6 @@ import {
 
 import { UserController } from '@src/apis/user/controllers/user.controller';
 import { UserResponseDto } from '@src/apis/user/dtos/response/user.response-dto';
-import { IdResponseDto } from '@src/libs/api/dtos/response/id.response-dto';
 import { HttpBadRequestException } from '@src/libs/exceptions/client-errors/exceptions/http-bad-request.exception';
 import { HttpConflictException } from '@src/libs/exceptions/client-errors/exceptions/http-conflict.exception';
 import { HttpForbiddenException } from '@src/libs/exceptions/client-errors/exceptions/http-forbidden.exception';
@@ -24,43 +22,6 @@ import {
 } from '@src/libs/types/type';
 
 export const ApiUser: ApiOperator<keyof Omit<UserController, 'verifyEmail'>> = {
-  Create: (
-    apiOperationOptions: ApiOperationOptionsWithSummary,
-  ): MethodDecorator => {
-    return applyDecorators(
-      ApiOperation({
-        ...apiOperationOptions,
-      }),
-      ApiCreatedResponse({
-        description: '정상적으로 회원가입 됨.',
-        type: IdResponseDto,
-      }),
-      HttpBadRequestException.swaggerBuilder(HttpStatus.BAD_REQUEST, [
-        {
-          code: COMMON_ERROR_CODE.INVALID_REQUEST_PARAMETER,
-          description:
-            '해당 필드는 request parameter 가 잘못된 경우에만 리턴됩니다.',
-          additionalErrors: {
-            errors: [
-              {
-                reason: 'reason',
-                property: 'property',
-                value: 'value',
-              },
-            ],
-            errorType: CustomValidationError,
-          },
-        },
-      ]),
-      HttpConflictException.swaggerBuilder(HttpStatus.CONFLICT, [
-        {
-          code: USER_ERROR_CODE.ALREADY_CREATED_USER,
-          description: '이미 존재하는 이메일입니다.',
-        },
-      ]),
-    );
-  },
-
   FindOne: (
     apiOperationOptions: ApiOperationOptionsWithSummary,
   ): MethodDecorator => {
