+++ src/libs/app-jwt/types/app-jwt.interface.ts
@@ -1,3 +1,3 @@
 export interface JwtPayload {
-  id: bigint;
+  sub: string;
 }
