+++ src/libs/exceptions/types/errors/auth/auth-error-code.constant.ts
@@ -1,6 +1,6 @@
 /**
  * 2000 ~ 2999
  */
-export const TOKEN_ERROR_CODE = {
+export const AUTH_ERROR_CODE = {
   WRONG_EMAIL_OR_PASSWORD: 2000,
 } as const;
