+++ src/libs/exceptions/types/errors/error-code.constant.ts
@@ -1,4 +1,4 @@
-import { TOKEN_ERROR_CODE } from '@src/libs/exceptions/types/errors/token/token-error-code.constant';
+import { AUTH_ERROR_CODE } from '@src/libs/exceptions/types/errors/auth/auth-error-code.constant';
 import { COMMON_ERROR_CODE } from '@src/libs/exceptions/types/errors/common/common-error-code.constant';
 import { USER_ERROR_CODE } from '@src/libs/exceptions/types/errors/user/user-error-code.constant';
 
@@ -8,5 +8,5 @@ export const ERROR_CODE = {
   // 1000 ~ 1999
   ...USER_ERROR_CODE,
   // 2000 ~ 2999
-  ...TOKEN_ERROR_CODE,
+  ...AUTH_ERROR_CODE,
 } as const;
