+++ src/libs/guards/providers/jwt-bearer-auth.strategy.ts
@@ -1,14 +1,10 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
-import { UserEntity } from '@src/apis/user/domain/user.entity';
-import { UserRepositoryPort } from '@src/apis/user/repositories/user.repository-port';
-import { USER_REPOSITORY_DI_TOKEN } from '@src/apis/user/tokens/di.token';
+import { JwtPayload } from '@src/libs/app-jwt/types/app-jwt.interface';
 import { ENV_KEY } from '@src/libs/core/app-config/constants/app-config.constant';
 import { AppConfigServicePort } from '@src/libs/core/app-config/services/app-config.service-port';
 import { APP_CONFIG_SERVICE_DI_TOKEN } from '@src/libs/core/app-config/tokens/app-config.di-token';
 import { Key } from '@src/libs/core/app-config/types/app-config.type';
-import { HttpUnauthorizedException } from '@src/libs/exceptions/client-errors/exceptions/http-unauthorized.exception';
-import { COMMON_ERROR_CODE } from '@src/libs/exceptions/types/errors/common/common-error-code.constant';
 import { ExtractJwt, Strategy } from 'passport-jwt';
 
 @Injectable()
@@ -16,27 +12,16 @@ export class JwtBearerAuthStrategy extends PassportStrategy(Strategy) {
   constructor(
     @Inject(APP_CONFIG_SERVICE_DI_TOKEN)
     private readonly appConfigService: AppConfigServicePort<Key>,
-    @Inject(USER_REPOSITORY_DI_TOKEN)
-    private readonly userRepository: UserRepositoryPort,
   ) {
     super({
       jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
       ignoreExpiration: false,
       secretOrKey: appConfigService.get<string>(ENV_KEY.JWT_SECRET),
+      issuer: appConfigService.get<string>(ENV_KEY.DOMAIN),
     });
   }
 
-  async validate(jwtPayload): Promise<UserEntity> {
-    const { payload } = jwtPayload;
-
-    const user = await this.userRepository.findOneById(payload.id);
-
-    if (!user) {
-      throw new HttpUnauthorizedException({
-        code: COMMON_ERROR_CODE.INVALID_TOKEN,
-      });
-    }
-
-    return user;
+  validate(payload: JwtPayload): JwtPayload {
+    return payload;
   }
 }
