+++ src/libs/interceptors/logging/request-response-logging.interceptor.ts
@@ -36,7 +36,7 @@ export class RequestResponseLoggingInterceptor implements NestInterceptor {
             method: method,
             url: originalUrl,
             body: request.body,
-            currentUser: request.user?.id,
+            currentUser: request.user?.sub,
             userAgent,
             ip,
           },
