+++ src/main.ts
@@ -1,7 +1,6 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import { BootstrapService } from '@src/bootstrap.service';
-import { AggregateID } from '@src/libs/ddd/entity.base';
 
 declare global {
   interface BigInt {
@@ -17,7 +16,9 @@ declare global {
   // eslint-disable-next-line @typescript-eslint/no-namespace
   namespace Express {
     export interface User {
-      id: AggregateID;
+      sub?: string;
+      email?: string;
+      password?: string;
     }
   }
 }
